GenAI/fastapi-simple/requirements.txt
fastapi[standard]
pydantic
openai
weaviate-client